feat: move pages to /app
- /offres/[typeOffre]/[idOffre]
- /offres/[typeOffre]/[idOffre]/partage
- /offres/[typeOffre]/partage-recherche
